+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneWindowsUpdateForBusinessFeatureUpdateProfileWindows10/MSFT_IntuneWindowsUpdateForBusinessFeatureUpdateProfileWindows10.psm1
@@ -22,6 +22,14 @@ function Get-TargetResource
         $FeatureUpdateVersion,
 
         [Parameter()]
+        [System.Boolean]
+        $InstallFeatureUpdatesOptional,
+
+        [Parameter()]
+        [System.Boolean]
+        $InstallLatestWindows10OnWindows11IneligibleDevice,
+
+        [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance]
         $RolloutSettings,
 
@@ -94,8 +102,11 @@ function Get-TargetResource
             if (-Not [string]::IsNullOrEmpty($DisplayName))
             {
                 $getValue = Get-MgBetaDeviceManagementWindowsFeatureUpdateProfile `
-                    -Filter "DisplayName eq '$DisplayName'" `
-                    -ErrorAction SilentlyContinue
+                    -All `
+                    -ErrorAction SilentlyContinue | Where-Object `
+                    -FilterScript {
+                        $_.DisplayName -eq $DisplayName
+                    }
             }
         }
         #endregion
@@ -126,21 +137,24 @@ function Get-TargetResource
 
         $results = @{
             #region resource generator code
-            Description           = $getValue.Description
-            DisplayName           = $getValue.DisplayName
-            FeatureUpdateVersion  = $getValue.FeatureUpdateVersion
-            RolloutSettings       = $complexRolloutSettings
-            Id                    = $getValue.Id
-            Ensure                = 'Present'
-            Credential            = $Credential
-            ApplicationId         = $ApplicationId
-            TenantId              = $TenantId
-            ApplicationSecret     = $ApplicationSecret
-            CertificateThumbprint = $CertificateThumbprint
-            Managedidentity       = $ManagedIdentity.IsPresent
-            AccessTokens          = $AccessTokens
+            Description                                       = $getValue.Description
+            DisplayName                                       = $getValue.DisplayName
+            FeatureUpdateVersion                              = $getValue.FeatureUpdateVersion
+            InstallFeatureUpdatesOptional                     = $getValue.InstallFeatureUpdatesOptional
+            InstallLatestWindows10OnWindows11IneligibleDevice = $getValue.InstallLatestWindows10OnWindows11IneligibleDevice
+            RolloutSettings                                   = $complexRolloutSettings
+            Id                                                = $getValue.Id
+            Ensure                                            = 'Present'
+            Credential                                        = $Credential
+            ApplicationId                                     = $ApplicationId
+            TenantId                                          = $TenantId
+            ApplicationSecret                                 = $ApplicationSecret
+            CertificateThumbprint                             = $CertificateThumbprint
+            ManagedIdentity                                   = $ManagedIdentity.IsPresent
+            AccessTokens                                      = $AccessTokens
             #endregion
         }
+
         $assignmentsValues = Get-MgBetaDeviceManagementWindowsFeatureUpdateProfileAssignment -WindowsFeatureUpdateProfileId $Id
         $assignmentResult = @()
         if ($assignmentsValues.Count -gt 0)
@@ -186,6 +200,14 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $FeatureUpdateVersion,
+
+        [Parameter()]
+        [System.Boolean]
+        $InstallFeatureUpdatesOptional,
+
+        [Parameter()]
+        [System.Boolean]
+        $InstallLatestWindows10OnWindows11IneligibleDevice,
 
         [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance]
@@ -242,43 +264,104 @@ function Set-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    $currentInstance = Get-TargetResource @PSBoundParameters
+    # Rollout is not "As soon as possible"
+    if ($null -ne $RolloutSettings)
+    {
+        # Rollout is either "On a specific date" or "gradually"
+        if (-not [string]::IsNullOrEmpty($RolloutSettings.OfferStartDateTimeInUTC))
+        {
+            $currentTime = Get-Date
+            $offerStartDate = [datetime]::Parse($RolloutSettings.OfferStartDateTimeInUTC)
+            # Rollout is "On a specific date"
+            if (-not [string]::IsNullOrEmpty($RolloutSettings.OfferEndDateTimeInUTC))
+            {
+                $offerEndDate = [datetime]::Parse($RolloutSettings.OfferEndDateTimeInUTC)
+                if ($offerEndDate -lt $offerStartDate.AddDays(1))
+                {
+                    throw 'OfferEndDateTimeInUTC must be greater than OfferStartDateTimeInUTC + 1 day.'
+                }
 
-    $PSBoundParameters.Remove('Ensure') | Out-Null
-    $PSBoundParameters.Remove('Credential') | Out-Null
-    $PSBoundParameters.Remove('ApplicationId') | Out-Null
-    $PSBoundParameters.Remove('ApplicationSecret') | Out-Null
-    $PSBoundParameters.Remove('TenantId') | Out-Null
-    $PSBoundParameters.Remove('CertificateThumbprint') | Out-Null
-    $PSBoundParameters.Remove('ManagedIdentity') | Out-Null
-    $PSBoundParameters.Remove('Verbose') | Out-Null
-    $PSBoundParameters.Remove('AccessTokens') | Out-Null
+                if ($offerEndDate -le $currentTime)
+                {
+                    throw 'OfferEndDateTimeInUTC must be greater than the current time.'
+                }
+            }
+        }
+    }
+
+    $currentInstance = Get-TargetResource @PSBoundParameters
+    $BoundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
+
+    if ($null -eq $RolloutSettings)
+    {
+        $BoundParameters.RolloutSettings = @{
+            OfferStartDateTimeInUTC = $null
+            OfferEndDateTimeInUTC = $null
+            OfferIntervalInDays = $null
+        }
+    }
 
     if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
     {
         Write-Verbose -Message "Creating an Intune Windows Update For Business Feature Update Profile for Windows10 with DisplayName {$DisplayName}"
-        $PSBoundParameters.Remove('Assignments') | Out-Null
+        $BoundParameters.Remove('Assignments') | Out-Null
 
-        $CreateParameters = ([Hashtable]$PSBoundParameters).clone()
+        if ($null -ne $RolloutSettings)
+        {
+            if (-not [string]::IsNullOrEmpty($RolloutSettings.OfferStartDateTimeInUTC))
+            {
+                $minTimeForAvailable = $currentTime
+                $newOfferStartDate = $offerStartDate
+                if ($offerStartDate -lt $minTimeForAvailable)
+                {
+                    $newOfferStartDate = $minTimeForAvailable
+                    $BoundParameters.RolloutSettings.OfferStartDateTimeInUTC = $newOfferStartDate.ToString('yyyy-MM-ddTHH:mm:ss.fffZ')
+                }
+
+                if (-not [string]::IsNullOrEmpty($RolloutSettings.OfferEndDateTimeInUTC))
+                {
+                    # If an end date is configured, then the start date must be at least the current time + 2 days
+                    $minTimeForAvailable = $currentTime.AddDays(2)
+                    if ($newOfferStartDate -lt $minTimeForAvailable)
+                    {
+                        Write-Verbose -Message 'OfferStartDateTimeInUTC must be at least the current time + 2 days, adjusting it...'
+                        $newOfferStartDate = $minTimeForAvailable
+                        $BoundParameters.RolloutSettings.OfferStartDateTimeInUTC = $newOfferStartDate.ToString('yyyy-MM-ddTHH:mm:ss.fffZ')
+                    }
+
+                    if ($offerEndDate -lt $newOfferStartDate.AddDays(1))
+                    {
+                        throw 'OfferEndDateTimeInUTC must be greater than OfferStartDateTimeInUTC + 1 day.'
+                    }
+
+                    if ($RolloutSettings.OfferIntervalInDays -gt ($offerEndDate - $newOfferStartDate).Days)
+                    {
+                        throw 'OfferIntervalInDays must be less than or equal to the difference between OfferEndDateTimeInUTC and OfferStartDateTimeInUTC in days.'
+                    }
+                }
+            }
+        }
+
+        $CreateParameters = ([Hashtable]$BoundParameters).Clone()
         $CreateParameters = Rename-M365DSCCimInstanceParameter -Properties $CreateParameters
         $CreateParameters.Remove('Id') | Out-Null
 
-        $keys = (([Hashtable]$CreateParameters).clone()).Keys
+        $keys = (([Hashtable]$CreateParameters).Clone()).Keys
         foreach ($key in $keys)
         {
-            if ($null -ne $CreateParameters.$key -and $CreateParameters.$key.getType().Name -like '*cimInstance*')
+            if ($null -ne $CreateParameters.$key -and $CreateParameters.$key.GetType().Name -like '*cimInstance*')
             {
                 $CreateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters.$key
             }
         }
-        #region resource generator code
-        $CreateParameters.Add("@odata.type", "#microsoft.graph.WindowsFeatureUpdateProfile")
-        $policy=New-MgBetaDeviceManagementWindowsFeatureUpdateProfile -BodyParameter $CreateParameters
-        $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $Assignments
 
-        if ($policy.id)
+        #region resource generator code
+        $policy = New-MgBetaDeviceManagementWindowsFeatureUpdateProfile -BodyParameter $CreateParameters
+
+        if ($policy.Id)
         {
-            Update-DeviceConfigurationPolicyAssignment -DeviceConfigurationPolicyId $policy.id `
+            $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $Assignments
+            Update-DeviceConfigurationPolicyAssignment -DeviceConfigurationPolicyId $policy.Id `
                 -Targets $assignmentsHash `
                 -Repository 'deviceManagement/windowsFeatureUpdateProfiles'
         }
@@ -287,24 +370,78 @@ function Set-TargetResource
     elseif ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
     {
         Write-Verbose -Message "Updating the Intune Windows Update For Business Feature Update Profile for Windows10 with Id {$($currentInstance.Id)}"
-        $PSBoundParameters.Remove('Assignments') | Out-Null
+        $BoundParameters.Remove('Assignments') | Out-Null
+        $BoundParameters.Remove('InstallLatestWindows10OnWindows11IneligibleDevice') | Out-Null
 
-        $UpdateParameters = ([Hashtable]$PSBoundParameters).clone()
+        if ($null -ne $RolloutSettings)
+        {
+            if (-not [string]::IsNullOrEmpty($RolloutSettings.OfferStartDateTimeInUTC))
+            {
+                $minTimeForAvailable = $currentTime
+                if (-not [string]::IsNullOrEmpty($currentInstance.RolloutSettings.OfferStartDateTimeInUTC))
+                {
+                    $currentOfferDate = [datetime]::Parse($currentInstance.RolloutSettings.OfferStartDateTimeInUTC)
+                }
+                else
+                {
+                    $currentOfferDate = [datetime]::MinValue
+                }
+                $newOfferStartDate = $offerStartDate
+
+                # If the configured start date is different from the current start date but less than the current time, we use the current start date
+                if ($offerStartDate -ne $currentOfferDate -and $offerStartDate -lt $minTimeForAvailable)
+                {
+                    if ($currentOfferDate -eq [datetime]::MinValue)
+                    {
+                        $newOfferStartDate = $minTimeForAvailable
+                        $currentOfferDate = $minTimeForAvailable
+                    }
+                    else
+                    {
+                        $newOfferStartDate = $currentOfferDate
+                    }
+                    $BoundParameters.RolloutSettings.OfferStartDateTimeInUTC = $newOfferStartDate.ToString('yyyy-MM-ddTHH:mm:ss.fffZ')
+                }
+
+                if (-not [string]::IsNullOrEmpty($RolloutSettings.OfferEndDateTimeInUTC))
+                {
+                    # If an end date is configured, then the start date must be at least the current time + 2 days
+                    $offerEndDate = [datetime]::Parse($RolloutSettings.OfferEndDateTimeInUTC)
+                    $minTimeForAvailable = $currentTime.AddDays(2)
+                    if ($newOfferStartDate -lt $minTimeForAvailable)
+                    {
+                        Write-Verbose -Message 'OfferStartDateTimeInUTC must be at least the current time + 2 days, adjusting it...'
+                        $newOfferStartDate = $minTimeForAvailable
+                        $BoundParameters.RolloutSettings.OfferStartDateTimeInUTC = $newOfferStartDate.ToString('yyyy-MM-ddTHH:mm:ss.fffZ')
+                    }
+
+                    if ($offerEndDate -lt $newOfferStartDate.AddDays(1))
+                    {
+                        throw 'OfferEndDateTimeInUTC must be greater than OfferStartDateTimeInUTC + 1 day.'
+                    }
+
+                    if ($RolloutSettings.OfferIntervalInDays -gt ($offerEndDate - $newOfferStartDate).Days)
+                    {
+                        throw 'OfferIntervalInDays must be less than or equal to the difference between OfferEndDateTimeInUTC and OfferStartDateTimeInUTC in days.'
+                    }
+                }
+            }
+        }
+
+        $UpdateParameters = ([Hashtable]$BoundParameters).Clone()
         $UpdateParameters = Rename-M365DSCCimInstanceParameter -Properties $UpdateParameters
-
         $UpdateParameters.Remove('Id') | Out-Null
 
-        $keys = (([Hashtable]$UpdateParameters).clone()).Keys
+        $keys = (([Hashtable]$UpdateParameters).Clone()).Keys
         foreach ($key in $keys)
         {
-            if ($null -ne $UpdateParameters.$key -and $UpdateParameters.$key.getType().Name -like '*cimInstance*')
+            if ($null -ne $UpdateParameters.$key -and $UpdateParameters.$key.GetType().Name -like '*cimInstance*')
             {
                 $UpdateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters.$key
             }
         }
 
         #region resource generator code
-        $UpdateParameters.Add("@odata.type", "#microsoft.graph.WindowsFeatureUpdateProfile")
         Update-MgBetaDeviceManagementWindowsFeatureUpdateProfile  `
             -WindowsFeatureUpdateProfileId $currentInstance.Id `
             -BodyParameter $UpdateParameters
@@ -346,6 +483,14 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $FeatureUpdateVersion,
+
+        [Parameter()]
+        [System.Boolean]
+        $InstallFeatureUpdatesOptional,
+
+        [Parameter()]
+        [System.Boolean]
+        $InstallLatestWindows10OnWindows11IneligibleDevice,
 
         [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance]
@@ -405,35 +550,120 @@ function Test-TargetResource
     Write-Verbose -Message "Testing configuration of the Intune Windows Update For Business Feature Update Profile for Windows10 with Id {$Id} and DisplayName {$DisplayName}"
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
-    $ValuesToCheck = ([Hashtable]$PSBoundParameters).clone()
-
-    if ($CurrentValues.Ensure -ne $Ensure)
-    {
-        Write-Verbose -Message "Test-TargetResource returned $false"
-        return $false
-    }
+    $ValuesToCheck = ([Hashtable]$PSBoundParameters).Clone()
     $testResult = $true
+
+    # Cannot be changed after creation
+    $ValuesToCheck.Remove('InstallLatestWindows10OnWindows11IneligibleDevice') | Out-Null
 
     #Compare Cim instances
     foreach ($key in $PSBoundParameters.Keys)
     {
+        if ($key -eq 'RolloutSettings')
+        {
+            continue
+        }
+
         $source = $PSBoundParameters.$key
         $target = $CurrentValues.$key
-        if ($source.getType().Name -like '*CimInstance*')
+        if ($null -ne $source -and $source.GetType().Name -like '*CimInstance*')
         {
             $testResult = Compare-M365DSCComplexObject `
                 -Source ($source) `
                 -Target ($target)
 
-            if (-Not $testResult)
+            if (-not $testResult)
             {
-                $testResult = $false
                 break
             }
 
             $ValuesToCheck.Remove($key) | Out-Null
         }
     }
+
+    if (-not $testResult)
+    {
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
+    if (($null -eq $RolloutSettings -and $null -ne $CurrentValues.RolloutSettings) -or `
+        ($null -ne $RolloutSettings -and $null -eq $CurrentValues.RolloutSettings))
+    {
+        Write-Verbose -Message 'RolloutSettings is null in either the desired configuration or the current configuration.'
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
+    $currentTime = Get-Date
+    [datetime]$offerStartDate = [datetime]::MinValue
+    [datetime]$offerEndDate = [datetime]::MinValue
+    [datetime]::TryParse($RolloutSettings.OfferStartDateTimeInUTC, [ref]$offerStartDate) | Out-Null
+    [datetime]::TryParse($RolloutSettings.OfferEndDateTimeInUTC, [ref]$offerEndDate) | Out-Null
+    if ($CurrentValues.Ensure -ne $Ensure)
+    {
+        if ($Ensure -eq 'Present')
+        {
+            if (($offerStartDate -ne [datetime]::MinValue -and $offerStartDate -lt $currentTime) `
+                    -and ($offerEndDate -ne [datetime]::MinValue -and $offerEndDate -lt $currentTime))
+            {
+                Write-Verbose -Message "Start and end time are in the past, skip the configuration."
+                Write-Verbose -Message "Test-TargetResource returned $true"
+                return $true
+            }
+        }
+
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
+    [datetime]$currentOfferStartDate = [datetime]::MinValue
+    [datetime]$currentOfferEndDate = [datetime]::MinValue
+    [datetime]::TryParse($CurrentValues.RolloutSettings.OfferStartDateTimeInUTC, [ref]$currentOfferStartDate) | Out-Null
+    [datetime]::TryParse($CurrentValues.RolloutSettings.OfferEndDateTimeInUTC, [ref]$currentOfferEndDate) | Out-Null
+    if (($offerEndDate -eq [datetime]::MinValue -and $currentOfferEndDate -ne [datetime]::MinValue) -or `
+        ($offerEndDate -ne [datetime]::MinValue -and $currentOfferEndDate -eq [datetime]::MinValue))
+    {
+        Write-Verbose -Message 'OfferEndDateTimeInUTC is null in either the desired configuration or the current configuration.'
+        Write-Verbose -Message "Test-TargetResource returned $false"
+        return $false
+    }
+
+    if ($offerStartDate -ne [datetime]::MinValue -and $currentOfferStartDate -ne [datetime]::MinValue)
+    {
+        if ($offerStartDate -ne $currentOfferStartDate `
+                -and $offerStartDate -gt $currentTime)
+        {
+            Write-Verbose -Message 'OfferStartDateTimeInUTC is different from current.'
+            $testResult = $false
+        }
+
+        if ($testResult -and $offerEndDate -ne [datetime]::MinValue -and $currentOfferEndDate -ne [datetime]::MinValue)
+        {
+            if ($offerStartDate -ne $currentOfferStartDate `
+                -and $offerStartDate -gt $currentTime `
+                -and $offerStartDate -lt $currentTime.AddDays(2))
+            {
+                Write-Verbose -Message 'OfferStartDateTimeInUTC must be greater than the current time + 2 days to be changable if OfferEndDateTimeInUTC is specified, resetting testResult to true.'
+                $testResult = $true
+            }
+
+            if ($offerEndDate -ne $currentOfferEndDate `
+                    -and $offerEndDate -gt $currentTime `
+                    -and $offerEndDate -gt $offerStartDate)
+            {
+                Write-Verbose -Message 'OfferEndDateTimeInUTC is different from current.'
+                $testResult = $false
+            }
+
+            if ($testResult -and $RolloutSettings.OfferIntervalInDays -ne $CurrentValues.RolloutSettings.OfferIntervalInDays)
+            {
+                Write-Verbose -Message 'OfferIntervalInDays is different from current.'
+                $testResult = $false
+            }
+        }
+    }
+    $ValuesToCheck.Remove('RolloutSettings') | Out-Null
     $ValuesToCheck.Remove('Id') | Out-Null
 
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
@@ -554,7 +784,7 @@ function Export-TargetResource
                 TenantId              = $TenantId
                 ApplicationSecret     = $ApplicationSecret
                 CertificateThumbprint = $CertificateThumbprint
-                Managedidentity       = $ManagedIdentity.IsPresent
+                ManagedIdentity       = $ManagedIdentity.IsPresent
                 AccessTokens          = $AccessTokens
             }
 
@@ -565,7 +795,7 @@ function Export-TargetResource
             {
                 $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString `
                     -ComplexObject $Results.RolloutSettings `
-                    -CIMInstanceName 'MicrosoftGraphwindowsUpdateRolloutSettings'
+                    -CIMInstanceName 'MicrosoftGraphWindowsUpdateRolloutSettings'
                 if (-Not [String]::IsNullOrWhiteSpace($complexTypeStringResult))
                 {
                     $Results.RolloutSettings = $complexTypeStringResult